place renderer in separate file
