+++ index.js
@@ -15,4 +15,5 @@ module.exports.octants = require('./lib/octants');
 module.exports.rasterize = require('./lib/rasterize');
 module.exports.nrrd = require('./lib/nrrd');
 module.exports.mesh = require('./lib/mesh');
+module.exports.renderer = require('./lib/renderer');
